Improve class existence check robustness, fixes #856
When the Magento installment contains class definitions that are
incomplete (e.g. a parent class extending from can not be auto-loaded),
a fatal error was given.

This fatal error is generally OK but on certain occasions it's wrong. One
such wrong occasion is when checking for rewrite conflicts in a class-
hierarchy based fashion.

Incomplete classes can be detected by checking via a terminating auto-
loader method for unresolvable based classes and then throwing an
exception.

Having a fatal-error wihtin Magento 1 for class_exists() is because of
the autoload implementation within Varien_Autoload. Without that
auto-loader, there is no fatal-error but just a warning in that case.

Both behaviors are covered by the testsuite.

Autoloader handling is provided via a new AutloadHandler utility.